requirements.txt
Flask==3.0.2
pandas==2.2.2
python-dotenv==1.0.1
schwab-py==1.0.1
openai==1.50.2
httpx==0.27.2
twilio==9.2.3
